minor refactor and clean
