+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/BridelleBelleMobileApplication.projitems
@@ -13,7 +13,7 @@
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Database\DocumnetDBClient.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ImageClient.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Database\ImageClient.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MagazineManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
@@ -24,6 +24,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Models\MagazineAdvertiser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Models\SocialMedia.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Models\Voucher.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Types\ImageType.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)MainPage.xaml">
