+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/MagazineManager.cs
@@ -4,27 +4,28 @@ using System.Text;
 using BridelleBelleMobileApplication.Database;
 using System.Threading.Tasks;
 using BridelleBelleMobileApplication.Models;
+using BridelleBelleMobileApplication.Types;
 
 namespace BridelleBelleMobileApplication
 {
-    public class MagazineManager
-    {
-	    private DocumnetDBClient docDb;
-        private ImageClient Client;
-	    public MagazineManager()
-	    {
-		    docDb = new DocumnetDBClient();
-            Client = new ImageClient();
-	    }
+	public class MagazineManager
+	{
+		private DocumnetDBClient docDb;
+		private ImageClient Client;
+		public MagazineManager()
+		{
+			docDb = new DocumnetDBClient();
+			Client = new ImageClient();
+		}
 
-	    public Task<Magazine> Get()
-	    {
-		    return docDb.Get("4");
-	    }
+		public Task<Magazine> Get()
+		{
+			return docDb.Get("4");
+		}
 
-        public async Task<string> GetCovers(string fileName)
-        {
-            return await Client.GetImages(fileName);
-        }
-    }
+		public async Task<string> GetCovers(string fileName)
+		{
+			return await Client.GetImages(ImageType.CoverImages, fileName);
+		}
+	}
 }
