+++ BridelleBelleMobileApplication/BridelleBelleMobileApplication/Types/ImageType.cs
@@ -1,0 +1,9 @@
+﻿namespace BridelleBelleMobileApplication.Types
+{
+	public enum ImageType
+	{
+		CoverImages,
+		Thumbnail,
+		MagazinePage
+	}
+}
